Ensure sign matches in comparisons

+++ src/main.cpp
@@ -7017,10 +7017,10 @@ bool static ProcessMessage(const CChainParams& chainparams, CNode* pfrom, string
             Misbehaving(pfrom->GetId(), 20);
             return error("message addr size() = %u", nNumItems);
         }
-        if (const auto nExpectedSize = (nNumItems * sizeof(CAddress)); nExpectedSize != vRecv.in_avail()) {
+        if (const auto nExpectedSize = (nNumItems * sizeof(CAddress)); nExpectedSize != static_cast<uint64_t>(vRecv.in_avail())) {
             LOCK(cs_main);
             Misbehaving(pfrom->GetId(), 20);
-            return error("malformed 'addr' payload. expected %u bytes, got %u instead", nExpectedSize, vRecv.in_avail());
+            return error("malformed 'addr' payload. expected %u bytes, got %u instead", nExpectedSize, static_cast<uint64_t>(vRecv.in_avail()));
         }
         vRecv.Rewind(GetSizeOfCompactSize(nNumItems));
 
@@ -7112,17 +7112,16 @@ bool static ProcessMessage(const CChainParams& chainparams, CNode* pfrom, string
 
     else if (strCommand == "inv")
     {
-
         const auto nNumItems = ReadCompactSize(vRecv);
         if (nNumItems > MAX_INV_SZ) {
             LOCK(cs_main);
             Misbehaving(pfrom->GetId(), 20);
             return error("message inv size() = %u", nNumItems);
         }
-        if (const auto nExpectedSize = (nNumItems * sizeof(CInv)); nExpectedSize != vRecv.in_avail()) {
+        if (const auto nExpectedSize = (nNumItems * sizeof(CInv)); nExpectedSize != static_cast<uint64_t>(vRecv.in_avail())) {
             LOCK(cs_main);
             Misbehaving(pfrom->GetId(), 20);
-            return error("malformed 'inv' payload. expected %u bytes, got %u instead", nExpectedSize, vRecv.in_avail());
+            return error("malformed 'inv' payload. expected %u bytes, got %u instead", nExpectedSize, static_cast<uint64_t>(vRecv.in_avail()));
         }
         vRecv.Rewind(GetSizeOfCompactSize(nNumItems));
 
@@ -7200,10 +7199,10 @@ bool static ProcessMessage(const CChainParams& chainparams, CNode* pfrom, string
             Misbehaving(pfrom->GetId(), 20);
             return error("message getdata size() = %u", nNumItems);
         }
-        if (const auto nExpectedSize = (nNumItems * sizeof(CInv)); nExpectedSize != vRecv.in_avail()) {
+        if (const auto nExpectedSize = (nNumItems * sizeof(CInv)); nExpectedSize != static_cast<uint64_t>(vRecv.in_avail())) {
             LOCK(cs_main);
             Misbehaving(pfrom->GetId(), 20);
-            return error("malformed 'getdata' payload. expected %u bytes, got %u instead", nExpectedSize, vRecv.in_avail());
+            return error("malformed 'getdata' payload. expected %u bytes, got %u instead", nExpectedSize, static_cast<uint64_t>(vRecv.in_avail()));
         }
         vRecv.Rewind(GetSizeOfCompactSize(nNumItems));
